Refactor indicators and strategies for improved compatibility with new C++ structure; update liveIG script for enhanced configuration options. Changes include: - Updated PositionManager for SL/TP calculations. - Refactored ATR and ATRLOG to use BasicCandle for initialization and updates. - Modified EMA, RSI, and STOCH to accept BasicCandle for calculations. - Removed unnecessary buffers in buy_heikin_green and sell_heikin_red strategies. - Adjusted liveIG script parameters for better trading strategy performance.

+++ cpp_strategies/src/PositionManager.hpp
@@ -23,7 +23,7 @@ public:
         if (config.use_atr_for_sl && current_atr > 0.0) {
             return calculateStopLossWithATR(config, current_atr, logger);
         } 
-        else if (config.use_minmax_for_sl && candle_manager.size() >= config.sl_minmax_periods) {
+        else if (config.use_minmax_for_sl && candle_manager.size() >= static_cast<size_t>(config.sl_minmax_periods)) {
             return calculateStopLossWithMinMax(
                 config, current_price, candle_manager, current_candle, is_long, logger);
         }
